The cart grant total was wrongly rounded because of a parseInt()

+++ view/frontend/web/js/view/payment/method-renderer/alma_monthly_payments.js
@@ -143,7 +143,7 @@ define(
             },
 
             totalPaid: function (cost) {
-                return priceUtils.formatPrice(parseInt(window.checkoutConfig.quoteData.grand_total) + (cost/100), window.checkoutConfig.priceFormat);
+                return priceUtils.formatPrice(parseFloat(window.checkoutConfig.quoteData.grand_total) + (cost/100), window.checkoutConfig.priceFormat);
             },
 
             getFeesMention: function (customerFee) {
